+++ src/lib/components/SecondBar.svelte
@@ -24,7 +24,7 @@
 	></div>
 {:else if prefs.secondStyle === 'top-bar'}
 	<div
-		class="pointer-events-none absolute top-0 left-0 h-1.5 transition-none"
+		class="pointer-events-none absolute top-0 left-0 h-12 transition-none"
 		style="width: {pct}%; background: {prefs.colorProgress}; opacity: {prefs.progressOpacity /
 			100};"
 	></div>